% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{$.statelist}
\alias{$.statelist}
\title{Extract time and budget as vector}
\usage{
\method{$}{statelist}(x, name)
}
\arguments{
\item{x}{a state list}

\item{name}{either \code{budget} or \code{time}}
}
\value{
a vector of the time or budgets
}
\description{
Convinience function to extract the time and budget.
}
\examples{
\dontrun{
result$budget     # result is a 'statelist' returned from simulate
result$time
}
}
\author{
Andreas Scheidegger
}
