% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.r
\name{costs.per.year}
\alias{costs.per.year}
\title{Calculate the total costs per year}
\usage{
costs.per.year(statelist, income)
}
\arguments{
\item{statelist}{a state list}

\item{income}{the same values as passed to
\code{simulate_network}. Either a scalar or vector.}
}
\value{
a vector of the total cost per year
}
\description{
The annual total costs are calculated. The total costs consist
of damage, failure and rehabilitation costs.
}
\author{
Andreas Scheidegger
}
