% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategies.r
\name{do.nothing}
\alias{do.nothing}
\title{rehabilitation strategy: no rehabilitation}
\usage{
do.nothing(state)
}
\arguments{
\item{state}{}
}
\value{
a state list
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Andreas Scheidegger
}
