% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.r
\name{expand}
\alias{expand}
\title{Model expansion of the network}
\usage{
expand(state, n.new, separat.budget = FALSE)
}
\arguments{
\item{state}{a state object}

\item{n.new}{\code{n.new} number of new pipes}

\item{separat.budget}{Boolan, if \code{TRUE} expansion cost are
not counted on the normal budget}
}
\value{
the expanded inventory
}
\description{
Expand the network with additional pipes. The diameter of these pipes is sampled.
}
\author{
Andreas Scheidegger
}
