% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deterioration.r
\name{fail}
\alias{fail}
\title{Model failures of the network}
\usage{
fail(inventory)
}
\arguments{
\item{inventory}{}
}
\value{
inventory with new failures
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Andreas Scheidegger
}
