% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.r
\name{failures.per.year}
\alias{failures.per.year}
\title{Calculate number of failures per year}
\usage{
failures.per.year(statelist)
}
\arguments{
\item{statelist}{a state list}
}
\value{
vector containing the number of failures per year
}
\description{
The number of failures per year is calculated
from a state list produced by \code{\link{simulate_network}}.
}
\author{
Andreas Scheidegger
}
