% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{initiate.network}
\alias{initiate.network}
\title{creates the initial network}
\usage{
initiate.network(inventory = 0, budget = Inf)
}
\arguments{
\item{inventory}{if \code{inventory} is an integer it specifies the
number of initial pipes, or alternatively it can be a \code{data.frame}
containing the initial inventory of pipes.}

\item{budget}{initial budget}
}
\value{
a state list
}
\description{
Creates the initial network that can be used with \code{\link{simulate_network}}.
}
\author{
Andreas Scheidegger
}
