% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.r
\name{make.empty.inventory}
\alias{make.empty.inventory}
\title{Initialize empty inventory}
\usage{
make.empty.inventory()
}
\value{
an inventory object
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Andreas Scheidegger
}
