% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.r
\name{pipes.built.per.year}
\alias{pipes.built.per.year}
\title{Calculate number of newly built pipes for each year}
\usage{
pipes.built.per.year(statelist)
}
\arguments{
\item{statelist}{a state list}
}
\value{
vector containing the number of newly built pipes for each year
}
\description{
The number of newly built pipes per year is calculated
from a state list produced by \code{\link{simulate_network}}.
}
\author{
Andreas Scheidegger
}
