% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.r
\name{pipes.inservice.per.year}
\alias{pipes.inservice.per.year}
\title{Returns the number of pipes in service for each year}
\usage{
pipes.inservice.per.year(statelist)
}
\arguments{
\item{statelist}{a state list}
}
\value{
vector containing the number of pipes in service
}
\author{
Andreas Scheidegger
}
