% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategies.r
\name{replace.more.failures.than}
\alias{replace.more.failures.than}
\title{Rehabilitation strategy: replace pipes with too many failures}
\usage{
replace.more.failures.than(state, max.failures)
}
\arguments{
\item{state}{a state list}

\item{max.failures}{maximal allowed number of failures}
}
\value{
a state list
}
\description{
Replace pipes with a high number of failures. Pipes are only
replaced if the budget remains positive.
}
\examples{
## define a strategy function that can be passed to simulate():
mystrategy <- . \%>\% replace.more.failures.than(max.failure=3)

## or combine multiple strategies to define more complex strategy:
mystrategy <- . \%>\%
replace.more.failures.than(max.failures=2) \%>\%
  replace.n.oldest(n=3) \%>\%
  replace.n.highest.risk(n=2, failure.rate=f.rate) \%>\%
  replace.older.than(max.age=8) \%>\%
  replace.n.random(n=4)

}
\author{
Andreas Scheidegger
}
