% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategies.r
\name{replace.n.highest.risk}
\alias{replace.n.highest.risk}
\title{Rehabilitation strategy: replace the \code{n} pipes with the highest risk}
\usage{
replace.n.highest.risk(state, n, failure.rate)
}
\arguments{
\item{state}{a state list}

\item{n}{maximal number of pipes to replace}

\item{failure.rate}{failur rate function. Typically the same as passed to \code{\link{simulate}}.}
}
\value{
a state list
}
\description{
Strategy to prioritize pipes with high risk. Pipes are only
replaced if the budget remains positive.
}
\details{
The Risk is defined as the product of the probability to fail next year
and the expected failure costs.
}
\examples{
## define a strategy function that can be passed to simulate():
mystrategy <- . \%>\% replace.n.highest.risk(n=2, failure.rate=f.rate)

## or combine multiple strategies to define more complex strategy:
mystrategy <- . \%>\%
replace.more.failures.than(max.failures=2) \%>\%
  replace.n.oldest(n=3) \%>\%
  replace.n.highest.risk(n=2, failure.rate=f.rate) \%>\%
  replace.older.than(max.age=8) \%>\%
  replace.n.random(n=4)

}
\author{
Andreas Scheidegger
}
