% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategies.r
\name{replace.n.random}
\alias{replace.n.random}
\title{Rehabilitation strategy: replace \code{n} randomly selected pipes}
\usage{
replace.n.random(state, n)
}
\arguments{
\item{state}{a state list}

\item{n}{maximal number of pipes to replace}
}
\value{
a state list
}
\description{
Replace a certain number of randomly chosen pipes. Pipes are only
replaced if the budget remains positive.
}
\examples{
## define a strategy function that can be passed to simulate():
mystrategy <- . \%>\% replace.n.random(n=10)

## or combine multiple strategies to define more complex strategy:
mystrategy <- . \%>\%
replace.more.failures.than(max.failures=2) \%>\%
  replace.n.oldest(n=3) \%>\%
  replace.n.highest.risk(n=2, failure.rate=f.rate) \%>\%
  replace.older.than(max.age=8) \%>\%
  replace.n.random(n=4)

}
\author{
Andreas Scheidegger
}
