% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategies.r
\name{replace.older.than}
\alias{replace.older.than}
\title{Rehabilitation strategy: replace pipes over \code{max.age}}
\usage{
replace.older.than(state, max.age)
}
\arguments{
\item{state}{a state list}

\item{max.age}{pipes older than max.age are replaced}
}
\value{
a state list
}
\description{
Strategy to replace pipes over a given age. Pipes are only
replaced if the budget remains positive.
}
\examples{
## define a strategy function that can be passed to simulate():
mystrategy <- . \%>\% replace.older.than(max.age=85)

## or combine multiple strategies to define more complex strategy:
mystrategy <- . \%>\%
replace.more.failures.than(max.failures=2) \%>\%
  replace.n.oldest(n=3) \%>\%
  replace.n.highest.risk(n=2, failure.rate=f.rate) \%>\%
  replace.older.than(max.age=8) \%>\%
  replace.n.random(n=4)

}
\author{
Andreas Scheidegger
}
