% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.r
\name{replacement.value}
\alias{replacement.value}
\title{Calculate replacement value}
\usage{
replacement.value(diameter)
}
\arguments{
\item{diameter}{diameter of the pipe [mm]}
}
\value{
replacement value in CHF
}
\description{
Based on eq(1) of "The water Network Management Chalange, Maurer (2017)",
assuming a pipe length of 100m
}
\author{
Andreas Scheidegger
}
