% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.r
\name{sample.diameter}
\alias{sample.diameter}
\title{Sample a diameter of a new pipe.}
\usage{
sample.diameter(n = 1)
}
\arguments{
\item{n}{number of samples}
}
\value{
a vector of diameters
}
\description{
The diameter distribution is based on a real data set from Lisa Scholten 
(pipe_data.csv). Refer to the old exercise on watermain break modelling
in the ETH Infrastructure Systems course by Max Maurer.
}
\author{
Andreas Scheidegger
}
