% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{simulate.network.period}
\alias{simulate.network.period}
\title{Simulate the network for a period of time}
\usage{
\method{simulate}{network.period}(statelist, n.years, expansion, rehabilitation,
  failure.rate, income = 0)
}
\arguments{
\item{statelist}{a state list}

\item{n.years}{number of years to simulate}

\item{expansion}{either a scalar describing the number of pipes added
every year to expand the pipe network, or a vector of length \code{n.years}.}

\item{rehabilitation}{a (combination of) rehabilitation strategy function(s). See details below.}

\item{failure.rate}{a function describing the probability of a pipe failing in the next year
given its age, number of previous failures, and the age at the last failure (if any).}

\item{income}{either a scalar describing the annual income, or a vector of length \code{n.years}.}
}
\value{
an updated state list
}
\description{
Simulates failures, expansion, rehabilitation, and costs of a water supply pipe network.
The simulation is stochastic.
}
\details{
The rehabilitation is defined by combining different simple replacement strategies.
See the example for how this can be done using the \code{mystrategy} function input.

The failure behavior is defined by the function \code{failure.rate}.
It calculates the probability of a pipe failing within the next year based on pipe age,
pipe age at the last failure, and the number of failures. Note, the model
makes the assumption that a pipe cannot fail more than once per year.

The costs are calculated as a function of the pipe diameter, assuming all pipes have a
length of 100 meters.
}
\seealso{
\code{\link{simulate.network}} provides a slightly more convenient interface.
}
\author{
Andreas Scheidegger
}
