% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{simulate_network}
\alias{simulate_network}
\title{Run a simulation}
\usage{
simulate_network(t.sim, expansion, rehabilitation, failure.rate, income = 0,
  initial.budget = Inf, initial.inventory = NULL)
}
\arguments{
\item{t.sim}{number of years to simulate}

\item{expansion}{either a scalar describing the number of pipes added
every year to expand the pipe network, or a vector of length \code{t.sim}.}

\item{rehabilitation}{a (combination of) rehabilitation strategy function(s). See details below.}

\item{failure.rate}{a function describing the probability of a pipe failing in the next year
given its age, number of previous failures, and the point in time of the last failure (if any).}

\item{income}{either a scalar describing the annual income, or a vector of length \code{t.sim}.}

\item{initial.budget}{initial budget}

\item{initial.inventory}{if \code{initial.inventory} equals \code{NULL} the simulation starts 
without pipes, if it is an integer it specifies the number of initial pipes,
or alternatively it can be a \code{data.frame} containing the initial inventory of pipes.}
}
\value{
A list of length \code{t.sim+1} containing all modeled states. A state
is a list consisting of the time, the budget, and the inventory at a given point in time.
}
\description{
Simulates failures, expansion, rehabilitation, and costs of a water supply pipe network.
The simulation is stochastic.
}
\details{
The rehabilitation is defined by combining different simple replacement strategies.
See the example for how this can be done using the \code{mystrategy} function input.

The failure behavior is defined with by the function \code{failure.rate}.
It calculates the probability, that a pipe fails within the next year based on pipe age,
pipe age at the last failure, and the number of failures. Note, the model
makes the assumption that a pipe cannot fail more than once per year.

The costs are calculated as a function of the pipe diameter, assuming all pipes have a
length of 100 meters.
}
\examples{
# Define the failure rate
f.rate <- function(age, age.last.failure, n.failure) {
  if(n.failure==0){
    return(1/30)
  } else {
    return(1/10)
  }
}


# Define a complicated (and probably useless) rehabilitation strategy
mystrategy <- . \%>\%
  replace.n.highest.risk(n=2, failure.rate=f.rate) \%>\%
  replace.more.failures.than(max.failures=5) \%>\%
  replace.older.than(max.age=100, max.cost=2e6) \%>\%
  replace.n.oldest(n=3) \%>\%
  replace.n.random(n=4)
# This defines a prioritized sequence of annual rehabilitation steps as follows: 
# each year, and as long as there is enough budget, replace first the 2 pipes 
# with the highest risk of failure, then all pipes with more than 5 failures,
# then all pipes more than 100 years old, then the 3 oldest remaining pipes, and
# finally replace 4 randomly selected pipes. Additionally, spendings on the 
# rehabilitation strategy replace.older.than can not exceed a maximum 
# budget of 2,000,000 CHF.

# Or, define a "do nothing" rehabilitation strategy (i.e. repairs only, no pipe replacement)
mystrategy <- . \%>\% do.nothing


# Run the simulation
result <- simulate_network(t.sim=100,          # run it for 100 years
                   expansion=0,                # do not expand the system
                   rehabilitation=mystrategy,  # use the strategy defined above
                   failure.rate=f.rate,        # use the failure rate defined above
                   income=1e6,                 # the annual income
                   initial.budget=30e6,        # the initial budget
                   initial.inventory=500)      # start the simulation with 500 new pipes

# View results
str(result)    # just a long list of states

# Convenient functions to extract budget or time are available
result$time  
result$budget

}
\seealso{
Different replacement strategies are implemented: \code{\link{replace.n.highest.risk}},
\code{\link{replace.n.oldest}}, \code{\link{replace.n.random}}, \code{\link{replace.older.than}},
\code{\link{replace.more.failures.than}}, \code{\link{do.nothing}}
}
\author{
Andreas Scheidegger
}
