\name{WaMaSim-package}
\alias{WaMaSim-package}
\alias{WaMaSim}
\docType{package}
\title{
\packageTitle{WaMaSim}
}
\description{
\packageDescription{WaMaSim}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{WaMaSim}
\packageIndices{WaMaSim}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{WaMaSim}

Maintainer: \packageMaintainer{WaMaSim}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
