% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{$.statelist}
\alias{$.statelist}
\title{Extract time or budget as vectors}
\usage{
\method{$}{statelist}(x, name)
}
\arguments{
\item{x}{a state list}

\item{name}{either \code{budget} or \code{time}}
}
\value{
a vector of the time or budget
}
\description{
Convenient functions to extract the time or budget.
}
\examples{
\dontrun{
result$budget     # result is a 'statelist' returned from simulate
result$time
}
}
\author{
Andreas Scheidegger
}
