% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{$.statelist}
\alias{$.statelist}
\title{Extract time or budget as vectors}
\usage{
\method{$}{statelist}(x, name)
}
\arguments{
\item{x}{a state list}

\item{name}{name of the element to extract}
}
\value{
a vector of the time or budget, or a state
}
\description{
Convenient functions to extract the time or budget.
}
\examples{
\dontrun{
str(result)      # result is a 'statelist' returned from simulate_network
result$budget    # vector of budget
result$time      # vector of time
result$time.22   # state list of time 22
}
}
\author{
Andreas Scheidegger
}
