% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategies.r
\name{do.nothing}
\alias{do.nothing}
\title{Rehabilitation strategy: no rehabilitation}
\usage{
do.nothing(state)
}
\arguments{
\item{state}{a state list}
}
\value{
a state list
}
\description{
Dummy strategy to model no rehabilitation at all.
}
\examples{
## define a strategy function that can be passed to simulate():
mystrategy <- . \%>\% do.nothing

}
\author{
Andreas Scheidegger
}
