% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategies.r
\name{do.nothing}
\alias{do.nothing}
\title{Rehabilitation strategy: no pipe replacement, repairs only}
\usage{
do.nothing(state)
}
\arguments{
\item{state}{a state list}
}
\value{
a state list
}
\description{
Dummy strategy to model no rehabilitation at all.
}
\examples{
## define a strategy function that can be passed to simulate_network():
mystrategy <- . \%>\% do.nothing

}
\seealso{
\code{\link{replace.n.highest.risk}},
\code{\link{replace.n.oldest}}, \code{\link{replace.n.random}}, \code{\link{replace.older.than}},
\code{\link{replace.more.failures.than}}
}
\author{
Andreas Scheidegger
}
