% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.r
\name{expand}
\alias{expand}
\title{Model expansion of the network}
\usage{
expand(inventory, n.new, time)
}
\arguments{
\item{inventory}{inventory object}

\item{n.new}{number of new pipes}

\item{time}{point in time of expansion}
}
\value{
the expanded inventory
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Andreas Scheidegger
}
