% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deterioration.r
\name{expand}
\alias{expand}
\title{Model expansion of the network}
\usage{
expand(inventory, n.new)
}
\arguments{
\item{inventory}{}

\item{n.new}{}
}
\value{
the expanded inventory
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Andreas Scheidegger
}
