% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.r
\name{fail}
\alias{fail}
\title{Model failures of the network}
\usage{
fail(inventory, failure.rate, time)
}
\arguments{
\item{inventory}{inventory object}

\item{failure.rate}{function returning the failure rate (age, time.last.failure, n.failure)}

\item{time}{point in time}
}
\value{
inventory with new failures
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Andreas Scheidegger
}
