% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.r
\name{fail}
\alias{fail}
\title{Model failures of the network}
\usage{
fail(state, failure.rate)
}
\arguments{
\item{state}{a state object}

\item{failure.rate}{function returning the failure rate.
\code{failure.rate} must take \code{age, time.last.failure, n.failure} as arguments.}
}
\value{
inventory with new failures
}
\description{
Lets pipe randomly fail. If a failure happened, the damage costs are
calculated and subtracted from the budget. Note, that this may result in a negative budget.
}
\author{
Andreas Scheidegger
}
