% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.r
\name{fail}
\alias{fail}
\title{Model failures of the network}
\usage{
fail(state, prob.failure)
}
\arguments{
\item{state}{a state object}

\item{prob.failure}{function returning the annual failure rate; i.e. the probability of a 
pipe failing in the current year of simulation.
\code{prob.failure} must take \code{age, age.last.failure, n.failure} as input arguments.
Note that in the case of a pipe with zero previous failures (i.e. \code{n.failure}=0), 
\code{age.last.failure}=NA.}
}
\value{
inventory with new failures
}
\description{
Allows pipes to randomly fail. If a failure occurs, the failure costs (repair + damage) are
calculated and subtracted from the budget. Note, that this may result in a negative budget.
}
\author{
Andreas Scheidegger
}
