% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.r
\name{failure.cost}
\alias{failure.cost}
\title{Calculate the (random) cost of a failure}
\usage{
failure.cost(diameter, mean = FALSE)
}
\arguments{
\item{diameter}{diameter [mm]}

\item{mean}{boolean. Should the expected cost be returned? Random otherwise.}
}
\value{
if \code{mean=FALSE}, the failure costs [CHF] are sampled
randomly. If \code{mean=TRUE}, the expected average costs are returned.
}
\description{
Calculate the costs caused by a failure according to Section 7.1 in
"The Water Network Management Challenge", Max Maurer 2017.
}
\author{
Andreas Scheidegger
}
