% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.r
\name{pipes.inservice.per.year}
\alias{pipes.inservice.per.year}
\title{Returns the number of pipes in service for each year}
\usage{
pipes.inservice.per.year(statelist)
}
\arguments{
\item{statelist}{a state list}
}
\value{
vector containing the number of pipes in service
}
\description{
The number of pipes in service is calculated for every year
based on a state list produced by \code{\link{simulate_network}}.
}
\author{
Andreas Scheidegger
}
