% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.r
\name{replacement.value}
\alias{replacement.value}
\title{Calculate replacement value}
\usage{
replacement.value(diameter)
}
\arguments{
\item{diameter}{diameter of the pipe [mm]}
}
\value{
replacement value [CHF]
}
\description{
Based on Eq(14) of "The Water Network Management Challenge, Max Maurer 2017",
assuming a pipe length of 100m.
}
\author{
Andreas Scheidegger
}
