% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\docType{package}
\name{WaMaSim}
\alias{WaMaSim}
\alias{WaMaSim-package}
\title{WaMaSim - Water Management Simulator}
\description{
WaMaSim is a package that simulates the effect of different
rehabilitation strategies for water distribution systems. It is an education tool
used for the Water Infrastructure Experimental and Computer Laboratory at ETH Zurich, Switzerland.
See the documentation for \code{\link{simulate_network}} to get started.
}
\author{
Andreas Scheidegger
}
